requirements.txt
Flask-Script==2.0.5
Flask-Session==0.3.1
jsonschema>=2.6.0
lupa==1.4